be/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg==3.2.7
psycopg-binary==3.2.7
qrcode==8.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
